worker farmer interaction test 1

+++ aquadSolution.c
@@ -93,7 +93,7 @@ double farmer(int numprocs) {
     // Assume at least as many tasks as workers
     for (i=0; i < numprocs; i++) {
         task = pop(tasks);
-        MPI_Send(&task, 5, MPI_INT, i+1, i, MPI_COMM_WORLD);
+        MPI_Send(&task, 5, MPI_DOUBLE, i+1, i, MPI_COMM_WORLD);
     }
     /*
     while (i<MAX_TASKS) {
